graph fixes

+++ schedules/getMoisture.js
@@ -9,7 +9,13 @@ module.exports.execute = (client) => {
 
 function getMoistureAndSave(client){
     moisture_data = client.helpers.arduinoBridge.getMoisture();
-    fs.writeFile('./moisture_data.csv', Date.now()+","+moisture_data+"\n", { flag: "a+" }, (err) => {
-    if (err) throw err;
-    }); 
+    if (fs.existsSync('./moisture_data.csv')) {
+        fs.writeFile('./moisture_data.csv', Date.now()+","+moisture_data+"\n", { flag: "a+" }, (err) => {
+        if (err) throw err;
+        }); 
+    } else {
+        fs.writeFile('./moisture_data.csv', "time,moisture"+"\n", { flag: "a+" }, (err) => {
+        if (err) throw err;
+        });
+    }
 }
